# openapi-orchestrate.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "OpsBob Backend",
                 "version":  "2.0.0"
             },
    "servers":  [
                    {
                        "url":  "https://opsbob-backend-4i6z7vcvfa-el.a.run.app",
                        "description":  "OpsBob Cloud Run backend"
                    }
                ],
    "paths":  {
                  "/webhook":  {
                                   "post":  {
                                                "summary":  "Receive Webhook",
                                                "description":  "Receives incident alerts from Instana or manual trigger.",
                                                "operationId":  "receive_webhook_webhook_post",
                                                "requestBody":  {
                                                                    "content":  {
                                                                                    "application/json":  {
                                                                                                             "schema":  {
                                                                                                                            "$ref":  "#/components/schemas/IncidentWebhook"
                                                                                                                        }
                                                                                                         }
                                                                                },
                                                                    "required":  true
                                                                },
                                                "responses":  {
                                                                  "200":  {
                                                                              "description":  "Successful Response",
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {

                                                                                                                                  }
                                                                                                                   }
                                                                                          }
                                                                          },
                                                                  "422":  {
                                                                              "description":  "Validation Error",
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                  }
                                                                                                                   }
                                                                                          }
                                                                          }
                                                              }
                                            }
                               },
                  "/stream/{incidentId}":  {
                                               "get":  {
                                                           "summary":  "Stream Analysis",
                                                           "description":  "Streams Bob\u0027s analysis + agent pipeline via SSE.",
                                                           "operationId":  "stream_analysis_stream__incidentId__get",
                                                           "parameters":  [
                                                                              {
                                                                                  "name":  "incidentId",
                                                                                  "in":  "path",
                                                                                  "required":  true,
                                                                                  "schema":  {
                                                                                                 "type":  "string",
                                                                                                 "title":  "Incidentid"
                                                                                             }
                                                                              }
                                                                          ],
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "Successful Response",
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {

                                                                                                                                             }
                                                                                                                              }
                                                                                                     }
                                                                                     },
                                                                             "422":  {
                                                                                         "description":  "Validation Error",
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                             }
                                                                                                                              }
                                                                                                     }
                                                                                     }
                                                                         }
                                                       }
                                           },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           }
                                   }
                   }
}